+++ dcpp/ThrottleManager.h
@@ -1,4 +1,4 @@
-/* 
+/*
  * Copyright (C) 2009-2010 Jacek Sieka, arnetheduck on gmail point com
  *
  * This program is free software; you can redistribute it and/or modify
@@ -43,7 +43,7 @@ namespace dcpp
 
 		/*
 		 * Throttles traffic and writes a packet to the network
-		 * Handle this a little bit differently than downloads due to OpenSSL stupidity 
+		 * Handle this a little bit differently than downloads due to OpenSSL stupidity
 		 */
 		int write(Socket* sock, void* buffer, size_t& len);
 
@@ -89,7 +89,7 @@ namespace dcpp
 		void waitToken();
 
 		// TimerManagerListener
-		void on(TimerManagerListener::Second, uint32_t /* aTick */) throw();
+		void on(TimerManagerListener::Second, uint64_t /* aTick */) throw();
 	};
 
 }	// namespace dcpp
