+++ dcpp/UploadManager.cpp
@@ -533,7 +533,7 @@ void UploadManager::reloadRestrictions(){
     limits.RenewList(NULL);
 }
 
-void UploadManager::on(TimerManagerListener::Minute, uint32_t /* aTick */) throw() {
+void UploadManager::on(TimerManagerListener::Minute, uint64_t /* aTick */) throw() {
     UserList disconnects;
     {
         Lock l(cs);
@@ -613,7 +613,7 @@ void UploadManager::on(AdcCommand::GFI, UserConnection* aSource, const AdcComman
 }
 
 // TimerManagerListener
-void UploadManager::on(TimerManagerListener::Second, uint32_t) throw() {
+void UploadManager::on(TimerManagerListener::Second, uint64_t) throw() {
     Lock l(cs);
     UploadList ticks;
