+++ eiskaltdcpp-gtk/src/searchspy.cc
@@ -665,7 +665,7 @@ void SearchSpy::on(ClientManagerListener::IncomingSearch, const string& s) throw
 	WulforManager::get()->dispatchGuiFunc(func);
 }
 
-void SearchSpy::on(TimerManagerListener::Minute, uint32_t tick) throw()
+void SearchSpy::on(TimerManagerListener::Minute, uint64_t tick) throw()
 {
 	typedef Func0<SearchSpy> F0;
 	F0 *func = new F0(this, &SearchSpy::updateFrameStatus_gui);
